+++ packages/fresh_dio/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.4
+
+- fix: hang on uncaught refresh token exception ([#120](https://github.com/felangel/fresh/pull/120))
+
+# 0.4.3
+
+- fix: clone `FormData` on retry
+- feat: support subclasses of `OAuth2Token` in `Fresh.oAuth2`
+- chore: upgrade to `mocktail ^1.0.0`
+
 # 0.4.2
 
 - feat: add `assert` to prevent infinite refresh loop